Improve readiness waiting in server tests
commands.TestServer waits for the Hugo server to be ready by calling:

```go
time.Sleep(2 * time.Second)

```

While this approach prevents us from needing to modify the server
command for tests, it also means that any new hugo server test would
add another two seconds to the test suite. The server is also not
guaranteed to be ready after two seconds.

commands.TestServer is currently the only test that runs the hugo
server command and waits for the server to be ready, but we will
eventually need more of these tests in order to address bugs when
running the hugo server command.

This change adds a "ready" channel to commands.serverCmd. The server
command sends to the channel when the server is ready for connections.
Tests can then block on receiving from the channel instead of sleeping
for a fixed interval.

This speeds up TestServer. Here are the results of running the test
with a clean test cache:

Sleeping two seconds:

```
$ go test github.com/gohugoio/hugo/commands -run TestServer
ok      github.com/gohugoio/hugo/commands       2.618s
```

Using a ready channel:

```
$ go test github.com/gohugoio/hugo/commands -run TestServer
ok      github.com/gohugoio/hugo/commands       0.556s
```

Fixes #9421