Changed wording a message
The message "Start building sites ..." (line 291) reads like a command to the user. Not usually an issue, but when it's taking a long time it looks like it's waiting for the user to do something.

Changed to "Starting to build sites ..." so it's clearer it's just a notification and requires no action.