youtube shortcode: Gen. query string w/ querify
The construction of the query string relied on a
generation method that assumed there would only
ever one be parameter (`autoplay`).

Adding support for an additional param e.g.
`start` requires detecting whether either option
was set in the shortcode, delimiting with `&` and
various other logic that are captured in
`querify`.

Move to using `querify` and ship, under separate
PR, addition of support for a `start` offset
parameter (see: #10521).