rss: Add config variable to output full content
The RSS 2.0 specification states:
"An item's description element holds character data that contains the
item's full content or a summary of its contents, a decision entirely
at the discretion of the publisher."

So far, we do the latter only. But it can sometimes be useful to have
the full content included instead, so let's add a config variable to
also support the former.