fix: dotpath for Hugo modules mounted as data
Before this change, data files from Hugo modules were always mounted at the
root of the `data` directory. The File and FileMetaInfo structs for modules
are different from 'native' data directories.

This changes how the keyParts for data files are generated so that data
from modules or native directories are treated the same.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>