content_map_page: Ensure homepage exists
Previously, it was possible to create an _index.md, but have it not be
built, such as by setting `draft: true`. This would result in there
being no homepage, and when e.g. a theme would call `.Site.Home`, this
would result in a “runtime error: invalid memory address or nil pointer
dereference”. Now, if the homepage does not end up being built, we fall
back to creating a default homepage, as if _index.md never exited.

Fixes #11441