docker: improve dart-sass installation step
The Docker build currently installs curl as a runtime dependency in the
final Dockerfile stage. The shell script to install the dart-sass is
also included in the final image. This change moves the dart-sass
download to a separate stage and copies the binary over to the final
stage.

It also fixes dart-sass for multi-platform builds by changing BUILDARCH
to TARGETARCH, ensuring that the dart-sass architecture matches the
output platform instead of the build platform.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>