Revert "Document PAGE.String"
This reverts commit bb30e7c08f471d2d65f7a81ebedc791b185e39ff.

We don't want to document this, at least not this way.

The PAGE.String method is currently meant to be used to give contextual information about a page (path, filename).

It's meant for error messages, so you can do

```
{{ warnf "foo not found in %q" .Page }}
```

But when documented as "Returns the absolute path to the file backing the given page", that is

1. Not entirely correct.
2. Makes it a formal API that people will use as a filename/path, and we can never change it (e.g. to add more contextual information)