markup/org: add custom org link resolve function for `denote`
The go-org library used by hugo to parse `.org` files now supports custom link
resolvers. This PR uses the new functionality to add support for
[denote](https://protesilaos.com/emacs/denote) style links.

This can be used to publish note libraries as-is using hugo.

I've written about the details in a blog post
[here](https://dominik.suess.wtf/posts/2025/01/02/denote-hugo/).

I don't know if this functionality is too specific to my usecase so I'm happy to
discuss alternatives which could be more modular/adaptable to other systems.