Always strip taxonomy name prefix from term name
If the taxonomy name differs when normalized versus unnormalized (e.g.,
because it has a space in it which is converted to a dash when
normalized), the plural tree key will not match as a prefix to the
term's base path (which is normalized). Use the non-normalized taxonomy
name instead to ensure that the prefix always matches and is trimmed.

Fixes: gohugoio/hugo#13422
Signed-off-by: Jason Cox <me@jasoncarloscox.com>