Add transform.PortableText
Using this in a content adapter could look like this:

```handlebars
{{ $projectID := "<myprojectid>" }}
{{ $url := printf "https://%s.api.sanity.io/v2021-06-07/data/query/production?query=*[0]"  $projectID }}
{{ $cacheKey := print $url (now.Format "2006-01-02T15:04") }}
{{ $opts := dict "key" $cacheKey }}
{{ $r := resources.GetRemote $url $opts }}
{{ $m := $r | transform.Unmarshal }}
{{ $result := $m.result }}
{{ $markdown := transform.PortableText $result.body }}
{{ $content := dict
  "mediaType" "text/markdown"
  "value" $markdown
}}
{{ $page := dict
  "content" $content
  "kind" "page"
  "path" $result.slug.current
  "title" $result.title
}}
{{ .AddPage $page }}
```

Note that this is still very rough around the edges, but I was pleasantly surprised that I could get something working right out of the gate.

Converting PortableText to Markdown and then back to HTML may seem sub optimal, but it fits right into Hugo's render hook etc. setup.