markup/goldmark: Add config to auto-prefix footnote IDs
This commit introduces a new configuration option, `enableAutoIDPrefix`, for the
footnote extension. When enabled, a unique prefix based on the document's ID
is prepended to footnote IDs, preventing ID clashes when multiple documents are
rendered together. For example, a footnote ID will change from `fn:1` to
`h-ae2bde661ce65c97-fn:1`.

This change also refactors the footnote extension configuration from a boolean to a
struct to accommodate this new option.

Closes #8045